OptionParser {optparse}	R Documentation

A function to create an instance of a parser object

Description

This function is used to create an instance of a parser object which when combined with the parse_args, make_option, and add_option methods is very useful for parsing options from the command line.

Usage

OptionParser(usage = "usage: %prog [options]", option_list = list(),
  add_help_option = TRUE, prog = NULL, description = "", epilogue = "")

Arguments

usage
The program usage message that will printed out if parse_args finds a help option, %prog is substituted with the value of the prog argument.

option_list
A list of of OptionParserOption instances that will define how parse_args reacts to command line options. OptionParserOption instances are usually created by make_option and can also be added to an existing OptionParser instance via the add_option function.

add_help_option
Whether a standard help option should be automatically added to the OptionParser instance.

prog
Program name to be substituted for %prog in the usage message (including description and epilogue if present), the default is to use the actual Rscript file name if called by an Rscript file and otherwise keep %prog.

description
Additional text for print_help to print out between usage statement and options statement

epilogue
Additional text for print_help to print out after the options statement

Value

An instance of the OptionParser class.

Author(s)

Trevor Davis.

References

Python's optparse library, which inspired this package, is described here: http://docs.python.org/library/optparse.html

See Also

parse_args make_option add_option

[Package optparse version 1.3.2 Index]