print_help {optparse}	R Documentation

Printing an usage message from an OptionParser object

Description

print_help print an usage message from an OptionParser object, usually called by parse_args when it encounters a help option.

Usage

print_help(object)

Arguments

object
A OptionParser instance.

Value

print_help uses the cat function to print out a usage message. It returns invisible(NULL).

Author(s)

Trevor Davis.

References

Python's optparse library, which inspired this package, is described here: http://docs.python.org/library/optparse.html

See Also

{parse_args} {OptionParser}

[Package optparse version 1.3.2 Index]