OptionParser-class {optparse}	R Documentation

Option Parser

Description

Option Parser

Slots

usage

The program usage message that will printed out if parse_args finds a help option, %prog is substituted with the value of the prog argument.

options

A list of of OptionParserOption instances that will define how parse_args reacts to command line options. OptionParserOption instances are usually created by make_option and can also be added to an existing OptionParser instance via the add_option function.

description

Additional text for print_help to print out between usage statement and options statement

epilogue

Additional text for print_help to print out after the options statement

Author(s)

Trevor Davis.

See Also

OptionParserOption

[Package optparse version 1.3.2 Index]