make_option {optparse}	R Documentation

Functions to enable our OptionParser to recognize specific command line options.

Description

add_option adds a option to a prexisting OptionParser instance whereas make_option is used to create a list of OptionParserOption instances that will be used in the option_list argument of the OptionParser function to create a new OptionParser instance.

Usage

make_option(opt_str, action = "store", type = NULL, dest = NULL,
  default = NULL, help = "", metavar = NULL)

add_option(object, opt_str, action = "store", type = NULL, dest = NULL,
  default = NULL, help = "", metavar = NULL)

Arguments

opt_str
A character vector containing the string of the desired long flag comprised of “–” followed by a letter and then a sequence of alphanumeric characters and optionally a string of the desired short flag comprised of the “-” followed by a letter.

action
A character string that describes the action optparse should take when it encounters an option, either “store”, “store_true”, or “store_false”. The default is “store” which signifies that optparse should store the specified following value if the option is found on the command string. “store_true” stores TRUE if the option is found and “store_false” stores FALSE if the option is found.

type
A character string that describes specifies which data type should be stored, either “logical”, “integer”, “double”, “complex”, or “character”. Default is “logical” if action %in% c("store_true", store_false), typeof(default) if action == "store" and default is not NULL and “character” if action == "store" and default is NULL. “numeric” will be converted to “double”.

dest
A character string that specifies what field in the list returned by parse_args should optparse store option values. Default is derived from the long flag in opt_str.

default
The default value optparse should use if it does not find the option on the command line. Default is derived from the long flag in opt_str.

help
A character string describing the option to be used by print_help in generating a usage message. %default will be substituted by the value of default.

metavar
A character string that stands in for the option argument when printing help text. Default is the value of dest.

object
An instance of the OptionParser class

Value

Both make_option and add_option return instances of class OptionParserOption.

Author(s)

Trevor Davis.

References

Python's optparse library, which inspires this package, is described here: http://docs.python.org/library/optparse.html

See Also

parse_args OptionParser

Examples

make_option("--longflag")
   make_option(c("-l", "--longflag"))
   make_option("--integer", type="integer", default=5)
   make_option("--integer", default=as.integer(5))  # same as previous

   # examples from package vignette
   make_option(c("-v", "--verbose"), action="store_true", default=TRUE,
       help="Print extra output [default]")
   make_option(c("-q", "--quietly"), action="store_false",
     dest="verbose", help="Print little output")
   make_option(c("-c", "--count"), type="integer", default=5,
       help="Number of random normals to generate [default %default]",
       metavar="number")
   make_option("--generator", default="rnorm",
       help = "Function to generate random deviates [default \"%default\"]")
   make_option("--mean", default=0,
       help="Mean if generator == \"rnorm\" [default %default]")
   make_option("--sd", default=1, metavar="standard deviation",
       help="Standard deviation if generator == \"rnorm\" [default %default]")

[Package optparse version 1.3.2 Index]